space_debris_website/backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
pandas==1.3.3
numpy==1.21.2
scikit-learn==1.4.2
joblib==1.1.0
plotly==5.3.1
gunicorn==20.1.0
python-dotenv==0.19.1
requests==2.26.0
pytest==6.2.5
matplotlib==3.4.3
seaborn==0.11.2
cesium==0.10.0
flask-socketio==5.3.0
pyjwt==2.4.0
tensorflow==2.8.0
redis==4.3.4
celery==5.2.7
prometheus-client==0.14.1 